pyproject.toml
[project]
name = "yao-demo-vehicle-app"
dynamic = ["version"]
description = "Vehicle CAN Data Dashboard - Real-time visualization and analysis of vehicle CAN bus data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
]

[dependency-groups]
dev = [
    "apx",
]

[tool.apx.metadata]
app-name = "yao-demo-vehicle-app"
app-slug = "yao_demo_vehicle_app"
app-module = "yao_demo_vehicle_app.backend.app:app"
metadata-path = "src/yao_demo_vehicle_app/_metadata.py"


[tool.uv.sources]
apx = { git = "https://github.com/databricks-solutions/apx.git" }

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
format-jinja = "{{version}}"

[tool.hatch.build.hooks.version]
path = "src/yao_demo_vehicle_app/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/yao_demo_vehicle_app/__dist__", "src/yao_demo_vehicle_app/_metadata.py"]
exclude = ["src/yao_demo_vehicle_app/ui"]
